+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [10.0.4] - 2025-07-09
+
+### Changed
+
+- Styling update
+- Show loading message for face detection
+- Update big smile prompt to make it clearer
+
 ## [10.0.3] - 2025-07-09
 
 ### Changed